requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
autopep8==1.5.7
azure-core==1.20.1
azure-storage-blob==12.9.0
brotlipy==0.7.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.6
django-admin-black==0.0.1
django-cors-headers==3.10.0
django-storages==1.12.3
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==3.2
isodate==0.6.0
itypes==1.2.0
Jinja2==3.0.2
MarkupSafe==2.0.1
msrest==0.6.21
oauthlib==3.1.1
olefile==0.46
Pillow==8.4.0
pip==21.2.4
pycodestyle==2.7.0
pycparser==2.21
PyJWT==2.1.0
pyOpenSSL==21.0.0
PyPDF2==1.26.0
PySocks==1.7.1
python-bidi==0.4.2
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
reportlab==3.5.67
requests==2.26.0
requests-oauthlib==1.3.0
setuptools==58.0.4
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
stripe==2.49.0
toml==0.10.2
twilio==7.0.0
typing-extensions==3.10.0.0
Unipath==1.1
uritemplate==4.1.1
urllib3==1.26.7
webencodings==0.5.1
wheel==0.37.0
whitenoise==5.3.0
xhtml2pdf==0.2.5
